refactor(6746): tighten command contract test support
