+++ crates/kamn-e2e-harness/tests/command_contract/external_execution_contract_tests/config_flag_mapping_contract_tests.rs
@@ -42,30 +42,23 @@ fn spec_c70_integration_config_flags_map_for_sdk_direct_external_enabled() {
 
 #[test]
 fn spec_c71_integration_config_flags_map_for_mcp_tau_external_enabled() {
-    let kolme_binary = temp_path("kolme-node-mcp-map");
-    write_stub_binary(&kolme_binary);
-    #[cfg(unix)]
-    set_executable(&kolme_binary);
-
-    let agent_binary = temp_path("agent-node-mcp-map");
-    write_stub_binary(&agent_binary);
-    #[cfg(unix)]
-    set_executable(&agent_binary);
-
-    let config = RunCommandConfig {
-        mode: "mcp-tau".to_owned(),
-        kolme_binary: kolme_binary.display().to_string(),
-        agent_binary: Some(agent_binary.display().to_string()),
-        external_execution: true,
-        evidence_dir: "/tmp/evidence".to_owned(),
-        scenario_ids: vec!["S-01".to_owned()],
-    };
-    let output = with_external_component_binaries(|| {
-        execute_run_contract(&config).expect("run output should render")
+    with_two_executable_binaries("kolme-node-mcp-map", "agent-node-mcp-map", |kolme, agent| {
+        let config = RunCommandConfig {
+            mode: "mcp-tau".to_owned(),
+            kolme_binary: kolme.display().to_string(),
+            agent_binary: Some(agent.display().to_string()),
+            external_execution: true,
+            evidence_dir: "/tmp/evidence".to_owned(),
+            scenario_ids: vec!["S-01".to_owned()],
+        };
+        let output = with_external_component_binaries(|| render_run(&config));
+        assert_contains_all(
+            &output,
+            &[
+                "\"integration_config\":",
+                "\"agent_binary_required\":true",
+                "\"external_execution_enabled\":true",
+            ],
+        );
     });
-    assert!(output.contains("\"integration_config\":"));
-    assert!(output.contains("\"agent_binary_required\":true"));
-    assert!(output.contains("\"external_execution_enabled\":true"));
-    let _ = std::fs::remove_file(kolme_binary);
-    let _ = std::fs::remove_file(agent_binary);
 }
