+++ crates/kamn-e2e-harness/tests/command_contract/integration_runtime_contract_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use super::shared_support::*;
 
 #[path = "integration_runtime_contract_tests/integration_config_contract_tests.rs"]
 mod integration_config_contract_tests;
