+++ crates/kamn-e2e-harness/tests/command_contract/parser_verify_contract_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use std::path::Path;
 
 #[test]
 fn spec_c01_parser_accepts_run_with_required_flags() {
@@ -86,34 +87,41 @@ fn spec_c07_verify_command_writes_deterministic_report_output() {
     let evidence_dir = temp_path("evidence");
     let output_path = temp_path("report.json");
     let chain_dump_path = temp_path("kolme_chain_dump.json");
-    std::fs::create_dir_all(&evidence_dir).expect("evidence dir should be created");
-    std::fs::write(
-        evidence_dir.join("manifest.json"),
-        r#"{"schema_version":"kamn.e2e.evidence-manifest.v3","run_id":"e2e-run","started_at":"2026-02-21T14:30:52Z","completed_at":"2026-02-21T14:35:12Z","duration_seconds":260,"execution_mode":"sdk-direct","infrastructure":{"kolme_version":"0.x.y","kamn_version":"0.1.0","kamn_commit":"49efe252","kamn_agent_lib_version":"0.1.0","agent_runtime":"sdk-direct","node_count":3,"agent_count":3,"storage_backend":"sqlite+postgres"},"scenarios":[],"summary":{"total_scenarios":15,"passed":13,"failed":1,"skipped":1,"kolme_blocks_produced":47,"messages_exchanged":128,"proofs_anchored":47,"proofs_verified":47}}"#,
-    )
-    .expect("manifest should be written");
-    std::fs::write(&chain_dump_path, valid_chain_dump_json())
-        .expect("chain dump should be written");
-
+    prepare_verify_fixture(&evidence_dir, &chain_dump_path);
     let config = VerifyCommandConfig {
         evidence_dir: evidence_dir.display().to_string(),
         kolme_chain_dump: chain_dump_path.display().to_string(),
         output: output_path.display().to_string(),
     };
-    let first = execute_verify_contract(&config).expect("verify should succeed");
-    let second = execute_verify_contract(&config).expect("verify should be deterministic");
+    assert_verify_report_is_deterministic(&config, &output_path);
+    cleanup_verify_fixture(&evidence_dir, &output_path, &chain_dump_path);
+}
+
+fn prepare_verify_fixture(evidence_dir: &Path, chain_dump_path: &Path) {
+    std::fs::create_dir_all(evidence_dir).expect("evidence dir should be created");
+    std::fs::write(evidence_dir.join("manifest.json"), verify_manifest_json())
+        .expect("manifest should be written");
+    std::fs::write(chain_dump_path, valid_chain_dump_json()).expect("chain dump should be written");
+}
+
+fn assert_verify_report_is_deterministic(config: &VerifyCommandConfig, output_path: &Path) {
+    let first = execute_verify_contract(config).expect("verify should succeed");
+    let second = execute_verify_contract(config).expect("verify should be deterministic");
+    let written = std::fs::read_to_string(output_path).expect("verification report should be written");
     assert_eq!(first, second);
-
-    let written =
-        std::fs::read_to_string(&output_path).expect("verification report should be written");
     assert_eq!(written, first);
-    assert!(written.contains("\"schema_check\""));
-    assert!(written.contains("\"proof_check\""));
-    assert!(written.contains("\"chain_check\""));
-    assert!(written.contains("\"content_check\""));
+    for marker in ["\"schema_check\"", "\"proof_check\"", "\"chain_check\"", "\"content_check\""] {
+        assert!(written.contains(marker));
+    }
+}
 
+fn cleanup_verify_fixture(evidence_dir: &Path, output_path: &Path, chain_dump_path: &Path) {
     let _ = std::fs::remove_file(output_path);
     let _ = std::fs::remove_file(chain_dump_path);
     let _ = std::fs::remove_file(evidence_dir.join("manifest.json"));
     let _ = std::fs::remove_dir(evidence_dir);
+}
+
+fn verify_manifest_json() -> &'static str {
+    r#"{"schema_version":"kamn.e2e.evidence-manifest.v3","run_id":"e2e-run","started_at":"2026-02-21T14:30:52Z","completed_at":"2026-02-21T14:35:12Z","duration_seconds":260,"execution_mode":"sdk-direct","infrastructure":{"kolme_version":"0.x.y","kamn_version":"0.1.0","kamn_commit":"49efe252","kamn_agent_lib_version":"0.1.0","agent_runtime":"sdk-direct","node_count":3,"agent_count":3,"storage_backend":"sqlite+postgres"},"scenarios":[],"summary":{"total_scenarios":15,"passed":13,"failed":1,"skipped":1,"kolme_blocks_produced":47,"messages_exchanged":128,"proofs_anchored":47,"proofs_verified":47}}"#
 }
