+++ crates/kamn-e2e-harness/tests/command_contract/phase_inventory_contract_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use super::shared_support::*;
 
 #[path = "phase_inventory_contract_tests/phase_progression_contract_tests.rs"]
 mod phase_progression_contract_tests;
