+++ crates/kamn-e2e-harness/tests/command_contract/scenario_evidence_contract_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use super::shared_support::*;
 
 #[path = "scenario_evidence_contract_tests/scenario_results_contract_tests.rs"]
 mod scenario_results_contract_tests;
